% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{edb_get_api_key}
\alias{edb_get_api_key}
\title{Get the current EleccionesDB API key}
\usage{
edb_get_api_key()
}
\value{
Cadena de texto con la clave, o NULL si no está definida.
}
\description{
Recupera la clave de API desde la opción, variable de entorno o entorno interno.
}
