% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{edb_get_base_url}
\alias{edb_get_base_url}
\title{Get the current EleccionesDB API base URL}
\usage{
edb_get_base_url()
}
\value{
Character string with the current base URL.
}
\description{
Get the current EleccionesDB API base URL
}
\examples{
edb_get_base_url()
}
