% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{edb_set_api_key}
\alias{edb_set_api_key}
\title{Set the EleccionesDB API key}
\usage{
edb_set_api_key(key, persist = FALSE)
}
\arguments{
\item{key}{Cadena de texto con la API key.}

\item{persist}{Lógico. Si \code{TRUE}, guarda la clave en tu \code{.Renviron} de usuario.}
}
\value{
Invisiblemente, la clave anterior (si existía).
}
\description{
Guarda la clave de API globalmente para todas las peticiones protegidas.
Por defecto, la clave se almacena en la opción de R y en la variable de entorno
para la sesión actual. Si \code{persist = TRUE}, también se añade a tu archivo \code{.Renviron}
para que esté disponible en futuras sesiones.
}
\examples{
\dontrun{
edb_set_api_key("TU_API_KEY")
}
}
